(refactor) use argument parser
