+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.6
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -21,6 +21,7 @@ let package = Package(
     dependencies: [
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
+		.package(url: "https://github.com/apple/swift-argument-parser", from: "1.0.0"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -28,9 +29,11 @@ let package = Package(
         .target(
             name: "SwiftyAcknowledgements",
             dependencies: []),
-		.target(
+		.executableTarget(
 			name: "GenerateLicenseFile",
-			dependencies: []),
+			dependencies: [
+				.product(name: "ArgumentParser", package: "swift-argument-parser"),
+			]),
 		.testTarget(
             name: "SwiftyAcknowledgementsTests",
             dependencies: ["SwiftyAcknowledgements"],
